// application/frontend/src/app/core/models/shipment-form-fields.ts
/**
 * @license
 * Copyright 2022 Google LLC
 *
 * Use of this source code is governed by an MIT-style
 * license that can be found in the LICENSE file or at
 * https://opensource.org/licenses/MIT.
 */

export enum ShipmentFormFields {
  Label = 'label',
  ShipmentType = 'shipment-type',
  PenaltyCost = 'penalty-cost',
  AllowedVehicles = 'allowed-vehicles',
  CostsPerVehicle = 'costs-per-vehicle',
  PickupToDeliveryLimits = 'pickup-to-delivery-limits',
  TimeLimit = 'time-limit',
  AbsoluteDetourLimit = 'absolute-detour-limit',
  RelativeDetourLimit = 'relative-detour-limit',
  LoadDemands = 'load-demands',
}